Virtualbox Vagrant boxes may be built
I've done this in a couple different places before, and each time I
re-invent the same basic script, heavily influenced from OpsCode's Bento

This provides a way to build a builder host that can be used with
vagrant. If the project would need to be deployed on to AWS, you may
consider using an AWS builder.